moved dependency injector
